requirements.txt
# Django Core
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Geospatial Support
django-geojson==4.0.0

# API & HTTP
requests==2.31.0
httpx==0.26.0
urllib3==2.1.0

# Scientific Computing & Calculations
numpy==1.26.3
scipy==1.12.0
pandas==2.2.0
astropy==6.0.0
skyfield==1.48

# Caching & Task Queue
redis==5.0.1
django-redis==5.4.0
celery==5.3.6
celery[redis]==5.3.6

# Environment Variables
python-decouple==3.8
python-dotenv==1.0.1

# Date & Time
python-dateutil==2.8.2
pytz==2024.1

# API Documentation
drf-spectacular==0.27.1
drf-yasg==1.21.7

# Authentication & Security
djangorestframework-simplejwt==5.3.1
django-allauth==0.60.1
PyJWT==2.8.0
cryptography==42.0.0

# Validation
marshmallow==3.20.2
pydantic==2.5.3

# AI & ML (for chatbot and predictions)
openai==1.10.0
langchain==0.1.6
transformers==4.37.0

# Data Processing
jsonschema==4.21.1
pyyaml==6.0.1
toml==0.10.2

# Image Processing (for visualization support)
Pillow==10.2.0

# Monitoring & Logging
sentry-sdk==1.40.0
django-extensions==3.2.3

# Testing
pytest==8.0.0
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
faker==22.5.0

# Code Quality
black==24.1.1
flake8==7.0.0
isort==5.13.2

# WSGI/ASGI Servers
gunicorn==21.2.0
uvicorn[standard]==0.27.0
daphne==4.1.0

# Utilities
python-slugify==8.0.2
django-model-utils==4.3.1
django-storages==1.14.2
boto3==1.34.34

# WebSocket Support (for real-time updates)
channels==4.0.0
channels-redis==4.1.0

# Rate Limiting
django-ratelimit==4.1.0

# Health Checks
django-health-check==3.18.1

